+++ MacPacker/CHANGELOG.md
@@ -2,6 +2,7 @@ v0.6
 - feat: hit space to open internal preview
 - feat: setting to change the breadcrumb position
 - feat: rar read support
+- feat: show folder / file icons in table
 - chore: add support for macOS 14
 
 v0.5